+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- JWT signing support for authentication and content exchange via interceptors.
+
 ## 0.5.0-rc3
 
 - Use clj-tools for